Correções devido a alteração da entrada
- DIstancia agora em float
- Vertices começam sempre do 1

+++ listaAdjacencia.h
@@ -9,11 +9,11 @@ typedef struct listaAdjacencia ListaAdjacencia;
 
 typedef struct grafo Grafo;
 
-Node* criaNode(int destino, int distancia, float velocidade);
+Node* criaNode(int destino, float distancia, float velocidade);
 
 Grafo* criaGrafo(int vertices);
 
-void insereAresta(Grafo* grafo, int origem, int destino, int distancia, float velocidade);
+void insereAresta(Grafo* grafo, int origem, int destino, float distancia, float velocidade);
 
 void imprimeGrafo(Grafo* grafo);
